+++ python/paddle/fluid/tests/unittests/test_slice_op.py
@@ -528,34 +528,5 @@ class TestSliceAPI(unittest.TestCase):
         assert np.array_equal(res_7, input[-1, 0:100, :, 2:-1])
 
 
-class TestImperativeVarBaseGetItem(unittest.TestCase):
-    def test_getitem_with_long(self):
-        with fluid.dygraph.guard():
-            data = np.random.random((2, 80, 16128)).astype('float32')
-            var = fluid.dygraph.to_variable(data)
-            sliced = var[:, 10:, :var.shape[1]]  # var.shape[1] is 80L here
-            self.assertEqual(sliced.shape, [2, 70, 80])
-
-            sliced = var[:, var.shape[0]:, var.shape[0]:var.shape[1]]
-            self.assertEqual(sliced.shape, [2, 78, 78])
-
-    def test_getitem_with_float(self):
-        def test_float_in_slice_item():
-            with fluid.dygraph.guard():
-                data = np.random.random((2, 80, 16128)).astype('float32')
-                var = fluid.dygraph.to_variable(data)
-                sliced = var[:, 1.1:, :var.shape[1]]
-
-        self.assertRaises(Exception, test_float_in_slice_item)
-
-        def test_float_in_index():
-            with fluid.dygraph.guard():
-                data = np.random.random((2, 80, 16128)).astype('float32')
-                var = fluid.dygraph.to_variable(data)
-                sliced = var[1.1]
-
-        self.assertRaises(Exception, test_float_in_index)
-
-
 if __name__ == '__main__':
     unittest.main()
